#1671: Move Swagger-related code to own package
- Swagger code moved to `de.symeda.sormas.rest.swagger` package
- Security information now included in Swagger specification
