+++ sormas-rest/pom.xml
@@ -95,6 +95,7 @@
 					<!-- Swagger Generation Config -->
 					<resourcePackages>
 						<resourcePackage>de.symeda.sormas.rest</resourcePackage>
+						<resourcePackage>de.symeda.sormas.rest.swagger</resourcePackage>
 					</resourcePackages>
 					<outputDirectory>${basedir}/target/</outputDirectory>
 					<outputFilename>swagger</outputFilename>
